+++ tensorly/contrib/sparse/cp_tensor.py
@@ -1,4 +1,6 @@
-from ...cp_tensor import cp_to_tensor, unfolding_dot_khatri_rao
+from ...cp_tensor import cp_to_tensor
+from ...tenalg import unfolding_dot_khatri_rao
+
 from .core import wrap
 from ...utils import DefineDeprecated
