+++ tensorly/decomposition/_constrained_cp.py
@@ -5,9 +5,11 @@ import tensorly as tl
 from ..random import random_cp
 from ._base_decomposition import DecompositionMixin
 from ..base import unfold
-from ..cp_tensor import CPTensor, unfolding_dot_khatri_rao, cp_norm, validate_cp_rank
+from ..cp_tensor import CPTensor, cp_norm, validate_cp_rank
 from ..tenalg.proximal import admm, proximal_operator, validate_constraints
 from ..tenalg.svd import svd_interface
+from ..tenalg import unfolding_dot_khatri_rao
+
 
 # Author: Jean Kossaifi
 #         Jeremy Cohen <jeremy.cohen@irisa.fr>
