+++ tensorly/tenalg/core_tenalg/__init__.py
@@ -6,10 +6,12 @@ from .outer_product import outer, batched_outer
 from .moments import higher_order_moment
 from ._tt_matrix import tt_matrix_to_tensor
 from ._batched_tensordot import tensordot
+from .mttkrp import unfolding_dot_khatri_rao
 
 from ..base_tenalg import TenalgBackend
 
 
+# Create the tensor algebra dispatching backend and register methods
 class CoreTenalgBackend(TenalgBackend, backend_name="core"):
     pass
 
@@ -24,3 +26,4 @@ CoreTenalgBackend.register_method("batched_outer", batched_outer)
 CoreTenalgBackend.register_method("higher_order_moment", higher_order_moment)
 CoreTenalgBackend.register_method("_tt_matrix_to_tensor", tt_matrix_to_tensor)
 CoreTenalgBackend.register_method("tensordot", tensordot)
+CoreTenalgBackend.register_method("unfolding_dot_khatri_rao", unfolding_dot_khatri_rao)
