+++ tensorly/tenalg/einsum_tenalg/__init__.py
@@ -6,6 +6,7 @@ from .outer_product import outer, batched_outer
 from .moments import higher_order_moment
 from ._tt_matrix import tt_matrix_to_tensor
 from ._batched_tensordot import tensordot
+from .mttkrp import unfolding_dot_khatri_rao
 
 from ..base_tenalg import TenalgBackend
 
@@ -24,3 +25,6 @@ EinsumTenalgBackend.register_method("batched_outer", batched_outer)
 EinsumTenalgBackend.register_method("higher_order_moment", higher_order_moment)
 EinsumTenalgBackend.register_method("_tt_matrix_to_tensor", tt_matrix_to_tensor)
 EinsumTenalgBackend.register_method("tensordot", tensordot)
+EinsumTenalgBackend.register_method(
+    "unfolding_dot_khatri_rao", unfolding_dot_khatri_rao
+)
